Ditch review section for anomaly
Instead of having a dedicated 'review' section (the part of the article
that displays the base anomaly), it's been integrated into the 'anomaly'
section, with the base anomaly being generated from the base anomaly
just the same as all the others, via diff patching against the reference
anomaly.

Additionally, the bit of code that constructs the HTML for an anomaly
has been moved to the function that changes section, to ensure that it
is called when switching to the base anomaly as well as when generating
a normal anomaly for the first time.